Move Certification unit test to Jest from Karma
